api: pool: introduce vector pool
Introduce the vector pool to hold the vector of a generic type such
as packets. Updated odp_pool_capability_t to advertise the new vector
pool capabilities and odp_pool_param_t to configure the vector pool.
The application shall use the max_vector_size parameter to configure
the max number of general types such as odp_packet_t that
can hold in the vector.

When vector pool used as odp_packet_t container,
either pktin subsystem can allocate a buffer to hold/form
the vector in packet input path as pktin producer or
the application can allocate the vector pool using
odp_packet_vector_alloc() from packet API to create a vector
of packets to store an array of packets as a vector.

Signed-off-by: Matias Elo <matias.elo@nokia.com>
Signed-off-by: Jerin Jacob <jerinj@marvell.com>